build(script): copy declarations .d.ts for bundle (#1306)
# Motivation

Neither esbuild nor tsc copy the `.d.ts` declarations (of Candid) to the
output folders. This means that in `@icp-sdk/canisters`, we need a
script to copy the `src/declarations/**/*.d.ts` files to the root
`./declarations` folder that gets bundled within the shipped library.

For convenience reasons and because, according to my experience, it's
useful to do so, we also copy the DID and IDL files. Those can be useful
for developers - at least, it has often happened that I used those
directly.

# Changes

- Add a new bash script to copy the declarations (similar to the script
used in oisy wallet signer originally copied from juno)
- Use script in `packages/canisters/package.json` scripts tag