chore(did): import did file to canisters starting with cmc (#1301)
# Motivation

We want to start migrating code from the legacy libraries to the new
multi-entry canisters lib. We also want to adopt a namespace pattern for
did files.

Aside from the downloaded did files there is also, and primarly, those
copied from the IC repo cloned next to the ic-js repo.

This script setup the copy of the cmc related did file to the new
library but, also setup the functions and changes that we can reuse to
copy the other did files.

# Changes

- Update `import-candid` to copy did files from IC repo to canisters'
declarations
- Use it for the cmc package