feat(canisters): add cmc declarations (#1316)
# Motivation

We want to embed the source code within `@icp-sdk/canisters`. Therefore,
we start with the CMC and its declarations.

# Notes

I generate the IC did with the job but it did not provide the new files
(e.g.
https://github.com/dfinity/icp-js-canisters/actions/runs/19859878610).
Maybe the job only updates files and do not add those to git. So I did
it on my machine using the same did file as in the `@dfinity/cmc` on
main.