feat(script): ic-mgmt tooling for canisters as source (#1333)
# Motivation

We want to embed code in `@icp-sdk/canisters/ic-management` and use this
library as main. Legacy `@dfinity/ic-management` will be converted to
re-export peer.

That's why the tooling needs to be updated for it.

# Changes

This is similar to what we have done for the cmc in:

- #1321 no agent update
- #1323 move build step of legacy
- #1322 no candid generation for legacy

To spare some effort, I packed all changes in this PR.