feat(script): tooling for canisters as source (#1376)
# Motivation

We want to embed code in `@icp-sdk/canisters/ledger/icp`, `icrc`, `nns`
and `sns` and use this library as main. Legacy libraries will be
converted to re-export peer.

That's why the tooling needs to be updated for it.

# Changes

This is similar to what we have done for the ic-mgmt in #1368