feat(ledger-icp): list re-export types for index (#1394)
# Motivation

We used to export everyting from `candid/index` but, we cannot do that
with namespace. Exporing `type *` won't resolve the issue therefore
let's cherry pick few types corresponding to the functions exposed by
the `IndexCanister` and hope it's enough for backwards compatibility.

Anyways, we recommend to start using the multi-entry library.

# Changes

- Cherry pick types