feat(canisters)!: rename IndexNgCanister into IndexCanister (#1400)
# Motivation

Follow-up of #1397. I propose to rename `IndexNgCanister`
into`IndexCanister`. The keyword "Ng" makes no sense for new developers
and as we discovered, all known Index canisters are "Ng" canisters.

# Notes

Since we removed the `IndexCanister` I think we can ship the same
breaking changes in `@dfinity/ledger-icrc`. This change is unrelated to
the did namespace but rather a choice of clean-up.

# Changes

- Remove keyowrd "Ng" from class and types