chore(release): set version for v89 (#1423)
# Motivation

We want to release the `assets` canister library that was integrated as
it into `@icp-sdk/canisters`. In future iteration, it's code source
should be reviewed for consistency with the code base and generally
speaking to align it on our standards but, nevertheless, we want to have
a starting point therefore the release.

# Changes

- Set version and add an entry in CHANGELOG