chore(tests): remove dependencies from integration (#1431)
# Motivation

I couldn't upgrade cleanly AgentJS in #1297 because the new integration
tests have explicit dependencies on the packages and the agent - i.e.
when I removed the AgentJS from the packages, the libraries were still
referenced in `package-lock` since it was used by the integration tests.

To avoid this issue and make the maintenance light, I propose to just
remove any dependencies from the integration packages. Both IDE and npm
seem clever enough to resolve implicitely the libraries present in the
workspace.

# Changes

- Remove explicit dependendices from `tests/integration/