feat(cycles): setup cycles ledger in @icp-sdk/canisters (#1440)
# Motivation

Surprisingly we never added the cycles ledger to the suite of JS
libraries. While it implements ICRC standards, it provides additional
features, notably `withdraw` (basically "topup with the ledger") which
I'll be using in Juno but, we might someday use in other projects or
even in OISY.

# Notes

The ledger does not seem to released that often, that is why, unlike
other libraries, I pinned the DID file import to a specific commit
deployed on mainnet. We can iterate from there.

# References

- Repo: https://github.com/dfinity/cycles-ledger
- Dashboard:
https://dashboard.internetcomputer.org/canister/um5iw-rqaaa-aaaaq-qaaba-cai

# Changes

- Add import did file in `import-candid`
- Generate IDL declarations
- Re-export DID and IDL from declaration entries
- Setup `ledger/cycles` in `@icp-sdk/canisters`
- List entry in README and add it to the docs generation as well
- Add new entry to size checker

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>